# Improve the performance of validate