@webNeat 2023-06-16
You are right, the `$report` change didn't make a difference in performance (at least in the test case I tried). So I reverted it.